code: optimize hair/eye color graphic algorithm
